Roll back electron to v6.1.7
This should fix the rejections we're getting from the Mac App Store.

## Background

Newer versions of the electron framework use a version of Chromium that calls private (undocumented) Apple APIs. Apple started rejecting apps that do that, and all users of the Electron SDK were affected.

As I understand it, the Electron team rolled back Chromium and/or created a patched Chromium version that disabled calls to those APIs, but this is a stopgap fix.

See:

- https://github.com/electron/electron/issues/20027
- https://github.com/electron/electron/pull/20970